Add support for `-y`, `-j <jobs>` and `-v` flags
These are used in conjunction with `opam depext -i` and are passed through to
the `opam install` command and allow non-interactive, parallel and verbose
output to be activated in the OPAM source builds.  This conveniently allows the
most common flags to `opam install` to be used by an equivalent
`opam depext -i` call.

Also fix the debug printing when using `opam depext -d` to flush output more
often and add newlines.