Check that packages are installable before choosing their depexts
This is useful, for example, to choose the right version of camlp4 on
opam 2.0, and avoid getting the latest, uninstallable system
version, that would bring wrong depexts.

NOTE: `opam list --installable` is known to blow up due to a bug in Dose
on older opam versions (e.g. 1.2.2) ; so this may be problematic on
older opams.