Support dep variables and 2.0.0~beta5 command-line format
This is not backwards compatible. The system analysis and generation of
the variable is done according to
https://github.com/ocaml/opam/issues/3058 (https://gist.github.com/AltGr/5bfc8cea6f01e74b95de79ceaba39369),
but is intended to be moved to opam at some point.

When that happens, this tool will be limited to handling the various
package managers, deciding from the opam variables using e.g. `opam var
os-distribution`.

Note that this drops support for "source" tags, which aren't supported
by beta5 anymore.