Support test-only and docs-only dependencies in opam files
This is added via `--with-test` and `--with-doc` to opam-depext.
Since depext is in maintenance mode, the support for env variables
was done without upgrading the vendored cmdliner.  opam 2.1 is due
to have direct support for depext, so there is no need to disrupt
this stable tree much more than the minimum required.

Fixes #114